requirements.txt
flet
paramiko
python-dotenv
